backend/config/settings.py
# -*- coding: utf-8 -*-
import os
from pathlib import Path
import environ

# 配置 PyMySQL 作为 MySQLdb 的替代
import pymysql
pymysql.install_as_MySQLdb()

# 初始化环境变量 
env = environ.Env()
environ.Env.read_env(os.path.join(Path(__file__).resolve().parent.parent.parent, '.env'))

# 项目根目录
BASE_DIR = Path(__file__).resolve().parent.parent.parent

# 安全密钥（从环境变量读取）
SECRET_KEY = env('SECRET_KEY', default='django-insecure-example-key-for-dev-only')

# 调试模式（生产环境需设为False）
DEBUG = env.bool('DEBUG', default=False)
ALLOWED_HOSTS = env.list('ALLOWED_HOSTS', default=['localhost', '127.0.0.1'])

# 应用配置
INSTALLED_APPS = [
    'django.contrib.admin',
    'django.contrib.auth',
    'django.contrib.contenttypes',
    'django.contrib.sessions',
    'django.contrib.messages',
    'django.contrib.staticfiles',
    # 修正应用注册路径，使用AppConfig类的完整路径
    'backend.apps.kg_visualize',
    # 'backend.apps.user_management',  # 暂时注释掉
    'corsheaders',
    'rest_framework',
]

MIDDLEWARE = [
    'django.middleware.security.SecurityMiddleware',
    'django.contrib.sessions.middleware.SessionMiddleware',
    'corsheaders.middleware.CorsMiddleware',  # 跨域中间件
    'django.middleware.common.CommonMiddleware',
    'django.middleware.csrf.CsrfViewMiddleware',
    'django.contrib.auth.middleware.AuthenticationMiddleware',
    'django.contrib.messages.middleware.MessageMiddleware',
    'django.middleware.clickjacking.XFrameOptionsMiddleware',
]

ROOT_URLCONF = 'config.urls'

# 模板配置
TEMPLATES = [
    {
        'BACKEND': 'django.template.backends.django.DjangoTemplates',
        'DIRS': [os.path.join(BASE_DIR, 'frontend')],  # 前端页面目录
        'APP_DIRS': True,
        'OPTIONS': {
            'context_processors': [
                'django.template.context_processors.debug',
                'django.template.context_processors.request',
                'django.contrib.auth.context_processors.auth',
                'django.contrib.messages.context_processors.messages',
            ],
        },
    },
]

WSGI_APPLICATION = 'config.wsgi.application'

# 数据库配置
DATABASES = {
    'default': env.db('DATABASE_URL', default=f'sqlite:///{BASE_DIR / "db.sqlite3"}')
}

# 密码验证
AUTH_PASSWORD_VALIDATORS = [
    {'NAME': 'django.contrib.auth.password_validation.UserAttributeSimilarityValidator'},
    {'NAME': 'django.contrib.auth.password_validation.MinimumLengthValidator'},
]

# 语言和时区
LANGUAGE_CODE = 'zh-hans'
TIME_ZONE = 'Asia/Shanghai'
USE_I18N = True
USE_TZ = True

# 静态文件配置
STATIC_URL = '/static/'
STATICFILES_DIRS = [os.path.join(BASE_DIR, 'frontend/static')]

# 默认主键字段
DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'

# 暂时使用Django默认用户模型
# AUTH_USER_MODEL = 'user_management.User'

# REST Framework 配置
REST_FRAMEWORK = {
    'DEFAULT_AUTHENTICATION_CLASSES': [
        'rest_framework.authentication.SessionAuthentication',
    ],
    'DEFAULT_PERMISSION_CLASSES': [
        'rest_framework.permissions.IsAuthenticated',
    ],
    'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
    'PAGE_SIZE': 20,
}

# 跨域配置（开发环境）
if DEBUG:
    CORS_ALLOW_ALL_ORIGINS = True  # 生产环境需指定具体域名
else:
    CORS_ALLOWED_ORIGINS = env.list('CORS_ALLOWED_ORIGINS', default=[])

# ChatGPT API 配置
CHATGPT_API_KEY = env('CHATGPT_API_KEY', default='123456789')
CHATGPT_API_URL = env('CHATGPT_API_URL', default='https://api.openai.com/v1/')
CHATGPT_BASE_URL = env('CHATGPT_BASE_URL', default='https://api.openai.com/v1/')
CHATGPT_MODEL = env('CHATGPT_MODEL', default='gpt-3.5-turbo')
CHATGPT_MAX_TOKENS = env.int('CHATGPT_MAX_TOKENS', default=300)
CHATGPT_TEMPERATURE = env.float('CHATGPT_TEMPERATURE', default=0.7)
CHATGPT_USE_OPENAI_LIB = env.bool('CHATGPT_USE_OPENAI_LIB', default=True)