requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
macholib==1.16.3
modulegraph==0.19.6
openai==1.88.0
packaging==25.0
py2app==0.28.6
pydantic==2.11.7
pydantic_core==2.33.2
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
PyMySQL==1.1.2
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
zipp==3.23.0
requests==2.32.4
PyMySQL==1.1.2
django-environ==0.11.2
openai==1.102.0
djangorestframework==3.14.0